Rename converter modules
