+++ cheetah/converters/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from cheetah.converters import astralavista, dontbmad, nocelot, nxtables
+from cheetah.converters import astra, bmad, nxtables, ocelot
