+++ cheetah/particles/particle_beam.py
@@ -571,7 +571,7 @@ class ParticleBeam(Beam):
     @classmethod
     def from_astra(cls, path: str, device=None, dtype=torch.float32) -> "ParticleBeam":
         """Load an Astra particle distribution as a Cheetah Beam."""
-        from cheetah.converters.astralavista import from_astrabeam
+        from cheetah.converters.astra import from_astrabeam
 
         particles, energy, particle_charges = from_astrabeam(path)
         particles_7d = torch.ones((particles.shape[0], 7))
